// src/types/common.ts
export interface CardCompany {
  name: string;
  color?: string;
}

export interface CardNumbers {
  first: string;
  second: string;
  third: string;
  fourth: string;
}

export interface CardExpiration {
  month: string;
  year: string;
}

export type CardName = string;

export type CardSecurityCode = string;

export interface CardPassword {
  first: string;
  second: string;
  third: string;
  fourth: string;
}

export interface LocalCard {
  numbers: CardNumbers;
  expiration: CardExpiration;
  name: CardName;
  nickname?: string;
}